Problem: impossible to timestamp events before publishing a command
This is useful for testing and disconnected systems

Solution: similarly to commands, event's timestamp will be overriden
only if it doesn't have one.

Also, this changes the behaviour of the repository. Before, the timestamp
of the repository was the timestamp of the last command. Now, it's the
timestamp of the last command's last event.