Problem: entities are forced to be mutable
Also, the requirement for entities and objects to have a default
empty constructors is onerous.

Solution: make layouts constructor+getters driven

Instead of defining layouts through a full set of accessors, a
single constructor drives the layout. Therefore a use of the builder
pattern and immutable (final) properties is now advised.

This solution changes some APIs:

1. `new Layout<>(klass)` is deprecated in favour of `Layout.forClass(klass)`