Problem: failure to start StandardRepository with duplicate indices
The problem occurs in OSGi environments when IndexEngine can (supposedly)
get re-wired to a new StandardRepository instance. If IndexEngine has the
index added, when StandardRepository is wired to events/commands again,
it will attempt to add those indices again.

Related issue: https://github.com/npgall/cqengine/issues/79

Solution: check if it was an attempt to add an equivalent index, and if it was,
swallow the exception.

This is not a perfect solution, but if https://github.com/npgall/cqengine/issues/79 gets
solved, it can be improved.