Problem: index definition syntax is slightly verbose
Most of the time, for most of use cases, there is no need for indices to be
aware of query options.

Also, having a two-arguments lambda prevents us from using
method references (which is a very common scenario when we need
to index a property as is)

Solution: make a one-argument functional interface a default
syntax in SimpleIndex and MultiValueIndex and add a #withQueryOptions
static helper to allow two-arguments functions to be supplied
to access query options.