Problem: Absence of local NTP server silently locks es4j up
The failure to connect to a local NTP server is hard to detect (nothing is
happening, repository is not starting). Since NTP is a UDP protocol, we can't
reliably tell if there's an NTP server listening. The only way to tell if we're
not getting a response.

Solution: throw an exception after some reasonably high delay (2s)

Fixes #157