Problem: min/max calculation is not most efficient for large sets
Solution: optimize support for min/max in PostgreSQL Navigable Index

A massive bug was unveiled during this development. PostgreSQL Navigable
Index used SerializableComparableAttribute wrapper, but those can't
be discovered in a HashMap because == check fails on #get(), since it's
not the same index that was originally passed. This led to CQEngine falling
back to the FallbackIndex (duh!) that's extremely slow as it requires
in-memory processing and doesn't use PostgreSQL-built indices. A workaround
for this bug has been applied but further improvements are highly welcome.
Chances are, SerializableComparableAttribute should be removed altogether
to remove the complexity associated with it.