Problem: a4d6711 changed HybridTimestamp navigable index key type
Previously, HybridTimestamp was serialized to long (BIGINT), it is
now serialized to BigInteger (NUMERIC). However, the version of the
index format wasn't incremented because format didn't change.

However, requiring everybody to manually drop indices with wrong
data type is not a great way to handle this case and this will surely
result in bug reports.

Solution: detect this case and drop "offending" indices automatically